make MakePaster len arg size_t

+++ src/mark.c
@@ -1248,7 +1248,7 @@ static int MarkScrollDownDisplay(int n)
 	return n;
 }
 
-void MakePaster(struct paster *pa, char *buf, int len, int bufiscopy)
+void MakePaster(struct paster *pa, char *buf, size_t len, int bufiscopy)
 {
 	FreePaster(pa);
 	pa->pa_pasteptr = buf;
